+++ test/sequel_test.rb
@@ -167,6 +167,20 @@ describe "Sequel model mixin" do
     assert_empty account.active_session_keys
   end
 
+  it "doesn't define password hash association without a password feature" do
+    # rodauth-omniauth enables the :login feature, but doesn't manage passwords
+    rodauth_class = Class.new(Rodauth::Auth)
+    rodauth_class.configure { enable :login }
+    Account.include Rodauth::Model(rodauth_class)
+
+    assert_nil Account.association_reflection(:password_hash)
+    refute Account.const_defined?(:PasswordHash, false)
+
+    account = Account.create(email: "user@example.com")
+    account.destroy # doesn't attempt to delete from account_password_hashes
+    refute account.exists?
+  end
+
   it "accepts passing association options hash" do
     account = build_account(association_options: { order: Sequel.desc(:created_at) }) do
       enable :active_sessions